+++ src/components/task.js
@@ -1,6 +1,6 @@
 import {monthNames} from '../const.js';
 import {formatTime} from '../utils/common.js';
-import AbstractComponent from './abstract-class.js';
+import AbstractComponent from './abstract-component.js';
 
 const createHashtagsMarkup = (hashtags) => {
   return hashtags
